Ecriture des textures, textureSampler, buffer etc. jusqu'avant les framebuffers

+++ src/S3DL/RenderWindow.cpp
@@ -27,7 +27,7 @@ namespace s3dl
             int width, height;
             glfwGetFramebufferSize(_window, &width, &height);
 
-            VkExtent2D actualExtent = { width, height };
+            VkExtent2D actualExtent = { (uint32_t) width, (uint32_t) height };
 
             actualExtent.width = std::max(capabilities.minImageExtent.width, std::min(capabilities.maxImageExtent.width, actualExtent.width));
             actualExtent.height = std::max(capabilities.minImageExtent.height, std::min(capabilities.maxImageExtent.height, actualExtent.height));
